PsychHID Linux/Windows: Improve ReceiveReports and error handling.
* ReceiveReports was also enqueuing completely empty reports in
  its buffers, leading to buffer overflow errors pretty quickly!
  Fixed.

* Error handling was spamming the console in an infinite loop if
  this happened: Now it does the proper things and is less chatty.

* locationId could provide a wrong value in certain setups. Fixed.

-> PsychHID for 32 bit Octave & Matlab on Linux rebuilt & tested.
-> 64-Bit Linux and Windows rebuilt pending.

git-svn-id: https://psychtoolbox-3.googlecode.com/svn/trunk@2436 f5efbe48-b203-0410-9429-fc5e098e5a54

+++ PsychSourceGL/Source/Common/PsychHID/PsychHID.h
@@ -99,7 +99,7 @@ struct recDevice
     char manufacturer[256];					// name of manufacturer
     char product[256];						// name of product
     char serial[256];						// serial number of specific product, can be assumed unique across specific product or specific vendor (not used often)
-    long locID;							// long representing location in USB (or other I/O) chain which device is pluged into, can identify specific device on machine
+    double locID;						// double representing location in USB (or other I/O) chain which device is pluged into, can identify specific device on machine
     long usage;							// usage page which defines general usage
     long usagePage;						// usage within above page which defines specific usage
     long interfaceId;                                           // HIDAPI only: USB interface id.
